1.实现初步的收藏功能； 2.Home页添加我的收藏； 2.完成tv库与torrent之间的简单通讯；

+++ libs/core/src/main/java/com/seiko/core/data/api/DanDanApiService.kt
@@ -34,7 +34,7 @@ internal interface DanDanApiService {
      * @param animeId 作品编号
      */
     @GET("api/v2/bangumi/{animeId}")
-    suspend fun getBangumiDetails(@Path("animeId") animeId: Int): BangumiDetailsResponse
+    suspend fun getBangumiDetails(@Path("animeId") animeId: Long): BangumiDetailsResponse
 
     /**
      * 获取用户近期关注但未看/未看完的番剧的列表。
